Module Cameras.Atlas_camera

Camera for the atlas screen

val atlas_camera : Camera.camera

Atlas screen camera position, y goes unused

val move_atlas_camera : int -> unit

Move edit camera according to delta

parameter dx
x delta

val pan_margin : int

Pixel margin that denotes when the user's mouse position at the edge of the screen should pan the screen

val pan_atlas_camera_if_needed : Tsdl.Sdl.window -> unit

Pan the atlas camera if the mouse is within pan_margin of the screen's edge

parameter window
Application's SDL window