Module Cameras.Edit_camera

Camera for the edit screen

type zoom =

| CloseZoom

| NormalZoom

| FarZoom

Levels of zoom

val zoom_level : zoom Stdlib.ref

Current zoom level

val view_width : unit -> int

Width in tiles of screen according to zoom

val view_height : unit -> int

Height in tiles of screen according to zoom

val edit_camera : Camera.camera

Edit screen camera position

val clamp_camera_to_bounds : unit -> unit

Ensure camera stays within bounds of world grid

val zoom_in : unit -> unit

Increase zoom level

val zoom_out : unit -> unit

Decrease zoom level

val move_edit_camera : int -> int -> unit

Move edit camera according to deltas

parameter dx
x delta

parameter dy
y delta

val get_edit_window_ui_w_h : Tsdl.Sdl.window -> int * (int * int)

Get the width and height of the edit window for tiles and the height of the edit UI window for the UI

parameter window
Application's SDL window

returns
Tuple of window width and window height and UI window height

val pan_margin : int

Pixel margin that denotes when the user's mouse position at the edge of the screen should pan the screen

val pan_edit_camera_if_needed : Tsdl.Sdl.window -> bool -> unit

Pan the edit camera if the mouse is within pan_margin of the screen's edge and the game is not paused for a popup

parameter window
Application's SDL window

parameter paused_for_popup