Module Rendering.Atlas

OpenGL program to draw the atlas screen

val quad_verts : (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

Triangle vertices that assemble to a quad to render onto

val atlas_vertex_shader_src : string

The atlas vertex shader

val atlas_fragment_shader_src : string

The atlas fragment shader

val setup_tile_buffers : offsets:(float * float) array -> colors:(float * float * float) array -> int

Set up tile buffers and return the VAO

parameter offsets
Array of (x, y) offsets for each tile

parameter colors
Array of (r, g, b) colors for each tile

returns
VAO bound to tile buffer

val render_tiles : ?fbo:Tsdl.Sdl.uint8 option -> Tsdl.Sdl.window -> int -> int -> int -> unit

Render tiles to the screen or a framebuffer

parameter fbo
If set, render to this framebuffer rather than the screen (used by Globe.render_globe_screen)

parameter window
Application's SDL window

parameter sprogram
Shader program to use

parameter vao
Vertex array object to draw

parameter num_instances
Number of tiles to draw

val create_render_texture : width:int -> height:int -> int * int

Create a framebuffer to render onto

parameter width
Width of framebuffer

parameter height
Height of framebuffer

returns
(fbo, tex) where fbo is the framebuffer and tex is the texture object

val make_tile_data : ?use_atlas_camera:bool -> unit -> (float * float) array * (float * float * float) array

Iterate over the world grid to populate offset and color arrays for each tile

returns
(offsets, colors)

val sprogram : int option Stdlib.ref

Shader program compiled for the current OpenGL context, or None

val render_atlas_screen : ?to_texture:bool -> Tsdl.Sdl.window -> int option

Render the atlas screen

parameter to_texture
Render to window if false, otherwise to a new texture

parameter window
Application's SDL window

returns
if to_texture then a texture else None