Module Rendering.Button

Logic for handling UI button rendering

type 'a ui_button = {

identifier : 'a;
(*
Assign an identifier to a button to map data defined in Rendering to functionality in Controls
*)

bounding_box : Tsdl.Sdl.rect;

texture_blob : Assets.Assetloader.asset_blob;

mutable initialized : bool;

}

UI button data

val ui_button_eq : 'a ui_button -> 'a ui_button -> bool

Check if two UI buttons are equal

val point_inside_button : (int * int) -> 'a ui_button -> bool

Check if a point is inside of a button

parameter x

parameter y

parameter button
The button to check in

returns
Whether (x, y) is within button.bounding_box

val any_buttons_uninit : 'a ui_button list -> bool

Check if any buttons are uninitialized in a list of buttons

parameter l
List to search over

returns
Whether any element of l satisfies not button.initialized

val render_buttons : Tsdl.Sdl.renderer -> 'a ui_button list -> unit

Render a list of buttons to the screen

parameter renderer
Application window's renderer

parameter l
List of buttons to render