Module Rendering.Edit

Rendering logic for the edit screen

val tile_texture_cache : (World.Biomes.biome_tile * int * int, Tsdl.Sdl.texture) Stdlib.Hashtbl.t

Cache for tile textures

val texture_of_tile : Tsdl.Sdl.renderer -> (World.Biomes.biome_tile * int * int) -> Tsdl.Sdl.texture

Look in cache for tile texture, otherwise render it

(biome, altitude, frame_count) is the key to tile_texture_cache

parameter renderer
Application window's SDL renderer

parameter biome
Biome of tile

parameter altitude
Altitude of tile

parameter frame_count
The current frame counter

returns
An SDL texture containing the rendered tile

val lifeform_texture_cache : (World.Life.Lifeform.lifeform * int, Tsdl.Sdl.texture) Stdlib.Hashtbl.t

Cache for lifeform textures

val texture_of_lifeform : Tsdl.Sdl.renderer -> (World.Life.Lifeform.lifeform * int) -> Tsdl.Sdl.texture

Look in cache for lifeform texture, otherwise render it

(lifeform, sprite_index) is the key to lifeform_texture_cache

parameter renderer
Application window's SDL renderer

parameter lf
The lifeform

parameter frame_count
The current frame counter

returns
An SDL texture containing the rendered lifeform

val render_edit_screen : Tsdl.Sdl.window -> int -> unit

Render the edit screen

parameter window
Application's SDL window

parameter frame_counter
The current frame counter