Module Rendering.Edit_screen

Rendering logic for the edit screen

val animated_tile_update_factor : int

How many frames it takes to draw the next frame of an animated tile

val tile_texture_cache : (World.Biomes.biome_tile * int * int, Tsdl.Sdl.texture) Stdlib.Hashtbl.t

Cache for tile textures

val texture_of_tile : Tsdl.Sdl.renderer -> (World.Biomes.biome_tile * int * int) -> Tsdl.Sdl.texture

Look in cache for tile texture, otherwise render it

(biome, altitude, frame_count) is the key to tile_texture_cache

parameter renderer
Application window's SDL renderer

parameter biome
Biome of tile

parameter altitude
Altitude of tile

parameter frame_count
The current frame counter

returns
An SDL texture containing the rendered tile

val lifeform_texture_cache : (World.Life.Lifeform.lifeform * int, Tsdl.Sdl.texture) Stdlib.Hashtbl.t

Cache for lifeform textures

val texture_of_lifeform : Tsdl.Sdl.renderer -> (World.Life.Lifeform.lifeform * int) -> Tsdl.Sdl.texture

Look in cache for lifeform texture, otherwise render it

(lifeform, sprite_index) is the key to lifeform_texture_cache

parameter renderer
Application window's SDL renderer

parameter lf
The lifeform

parameter frame_count
The current frame counter

returns
An SDL texture containing the rendered lifeform

val ui_bg_color : int * int * int

val ui_bevel_light_color : int * int * int

val ui_bevel_dark_color : int * int * int

val black_color : int * int * int

type edit_button =

| Volcano

| Asteroid

val edit_screen_buttons : edit_button Ui_button.ui_button list Stdlib.ref

val selected_edit_screen_button : int Stdlib.ref

val init_edit_screen_buttons : (int -> int -> int * int) -> (int -> int) -> (int -> int) -> int -> unit

val draw_edit_ui : Tsdl.Sdl.window -> Tsdl.Sdl.renderer -> (int * int) -> unit

Draw the edit screen UI

parameter window
Application's SDL window

parameter renderer
Application window's SDL renderer

parameter cursor_pos
Screen-space position of cursor

val render_edit_screen : Tsdl.Sdl.window -> int -> (int * int) -> unit

Render the edit screen

parameter window
Application's SDL window

parameter frame_counter
The current frame counter

parameter cursor_pos
Screen-space position of the cursor