Module Rendering.Edit_ui

Rendering logic for the Edit screen UI

type edit_button =

| Volcano

| Asteroid

| Examine

Identifiers for edit_screen_buttons

val edit_screen_buttons : edit_button Button.ui_button list Stdlib.ref

List of buttons to be drawn on the edit screen

val selected_edit_screen_button : int Stdlib.ref

Index into edit_screen_buttons of selected edit screen button

val init_edit_screen_buttons : (int -> int -> int * int) -> (int -> int) -> (int -> int) -> int -> unit

Initialize the edit screen buttons' positions and sizes

parameter pos
Function from (row, col) to (x, y)

parameter w
Function from col to w

parameter h
Function from col to h

parameter base_col
The first column of the edit screen to render in

val ui_bevel_w : int Stdlib.ref

Width of the main UI window's bevel

val draw_edit_ui : Tsdl.Sdl.window -> Tsdl.Sdl.renderer -> (int * int) -> unit

Draw the edit screen UI

parameter window
Application's SDL window

parameter renderer
Application window's SDL renderer

parameter cursor_pos
Screen-space position of cursor

val draw_popups : Tsdl.Sdl.window -> Tsdl.Sdl.renderer -> int -> unit

Draw any applicable popups

parameter window
The application's SDL window

parameter renderer
The applciation window's SDL renderer

parameter frame_count
The current frame count

val render_edit_ui : Tsdl.Sdl.window -> Tsdl.Sdl.renderer -> (int * int) -> int -> unit

Render the edit UI onto a texture, then copy it onto the screen (mostly just an exercise, perf is fine)

parameter window
The application's SDL window

parameter renderer
The application window's SDL renderer

parameter cursor_pos
Screen-space position of cursor

parameter frame_count
The current frame count