Module Rendering.Globe_data

Data for the globe screen

val rotation_lat : float Stdlib.ref

Latitudinal rotation position of globe in degrees

val rotation_lon : float Stdlib.ref

Longitudinal rotation position of globe in degrees

val velocity_lat : float Stdlib.ref

Latitudinal velocity of globe in degrees per frame

val min_abs_velocity_lon : float

Minimum longitudinal velocity of globe in degrees per frame

val velocity_lon : float Stdlib.ref

Longitudinal velocity of globe in degrees per frame

val max_globe_speed : float

Maximum globe spin speed in degrees per frame

val globe_spin_friction : float

Decay of rotational velocity in degrees per frame per frame

val globe_pinned : bool Stdlib.ref

Whether the globe is being held in place by the user

val globe_last_mouse_pos : (int * int) option Stdlib.ref

Last mouse recorded position of the mouse while pinning the globe