Module Rendering.Globe_screen_opengl

OpenGL program to draw the globe screen

val quad_vertices : (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

Triangle vertices that assemble to a quad to render onto, packed with UVs

val globe_vertex_shader_src : string

The globe vertex shader

val globe_fragment_shader_src : string

The globe fragment shader

val setup_fullscreen_quad : unit -> int

Set up fullscreen quad buffer and return the VAO

returns
VAO bound to quad buffer

val sprogram : int option Stdlib.ref

Shader program compiled for the current OpenGL context, or None

val atlastex : int option Stdlib.ref

Cached atlas texture, or None

val render_globe_screen : Tsdl.Sdl.window -> int -> unit

Render the globe screen

parameter window
Application's SDL window

parameter frame_counter
The current frame counter