Module Rendering.Popup

Facilities for drawing popup windows

type popup = {

bounding_box : Tsdl.Sdl.rect;

initialized : bool;

}

val pause_everything_for_popup : bool Stdlib.ref

Whether to pause simulation and input to render popup and wait for user to click out

val draw_popup : ?bg_color:(int * int * int) -> ?dark_bevel_color:(int * int * int) -> ?light_bevel_color:(int * int * int) -> ?bevel_w:int option -> Tsdl.Sdl.renderer -> popup -> int * int * ((int * int) * (int * int))

Draw a popup

parameter bg_color
Background color

parameter dark_bevel_color
Color of the bottom/right edges of the bevel

parameter light_bevel_color
Color of the top/left edges of the bevel

parameter bevel_w
Thickness of the bevel if set, otherwise auto-computed based on w

parameter renderer
Application window's SDL renderer

parameter x

parameter y

parameter w

parameter h

returns
(bevel_w, ui_buffer, (ui_position, ui_dimensions)) where ui_buffer is a computed value indicating suggested distance between UI area boundaries and UI elements

val point_inside_popup : (int * int) -> popup -> bool

Check if a point is inside a popup

parameter x

parameter y

parameter p
Popup to check inside of