Module Rendering.Text

Text rendering utilities via SDL-ttf

type font_family =

| CourierPrime

| Jacquard12

| NewPortLand

Font families included in build

type text_style =

| Regular

| Bold

| Italic

| BoldItalic

val font_blob_of_font_style : font_family -> text_style -> Assets.Assetloader.asset_blob

Get a font's asset_blob given its family and style

val text_cache : (font_family * text_style * int * (int * int * int) * string, Tsdl.Sdl.texture * Tsdl_ttf.Ttf.font) Stdlib.Hashtbl.t

A cache for rendered string textures

val clear_text_cache : unit -> unit

Clear text_cache and destroy its textures

val render_text : ?font_family:font_family -> ?style:text_style -> ?ptsize:int -> ?color:(int * int * int) -> ?fit:(int * int) -> ?draw_bounding_box:bool -> Tsdl.Sdl.renderer -> (int * int) -> ('a, unit, string, int * int) Stdlib.format4 -> 'a

Render a formatted string to the screen

parameter font_family

parameter style

parameter ptsize
Point size of text to render

parameter color
Color of text to render

parameter fit
If set, defines a bounding box that ptsize will be scaled up to match

parameter draw_bounding_box
For debugging, indicates to draw a red bounding box around the rendered texture

parameter renderer
Application's SDL renderer

parameter x

parameter y

parameter fmt
Format string

parameter ...
Format string arguments

returns
The height and width of the rendered texture, clipping invisiblie padding caused by a font's ascent and descent