Module Utils.Globals

Global data and pseudo-data

val tile_w : int Stdlib.ref

Width and height of tiles on the edit screen in pixels

val tile_h : int Stdlib.ref

Width and height of tiles on the edit screen in pixels

val zoom_factor : float Stdlib.ref

Size factor for scaled tile width and heigth

val scaled_tile_w : unit -> int

Tile width scaled w.r.t. zoom_factor

val scaled_tile_h : unit -> int

Tile height scaled w.r.t. zoom_factor

val need_to_flush_edit_caches : bool Stdlib.ref

Whether the cache of tile textures needs to be flushed

val clear_edit_cache : unit -> unit

Signals to flush the tile texture cache