Module Utils.Logging

logging.ml - Custom logging and error messages

type log_type =

| Log_None

| Log_Debug

| Log_Info

| Log_Warning

| Log_Error

| Log_Critical

Represents the severity of a log statement

val _GLOBAL_LOG_LEVEL : log_type

val int_of_log : log_type -> int

Follows the order in the type definition, [0:5]

type return_code = int * string

For exits, their appropriate return code and the message type

val rc_Ok : int * string

val rc_Error : int * string

val rc_SDL : int * string

val rc_OpenGL : int * string

val ansi_bold : string

ANSI encoding for bold text

val ansi_red : string

ANSI encoding for red text

val ansi_orange : string

ANSI encoding for orange text

val ansi_yellow : string

ANSI encoding for yellow text

val ansi_reset : string

ANSI encoding for plain text

val error_red : string

ANSI encoding for bold red text

val error_orange : string

ANSI encoding for bold orange text

val error_yellow : string

ANSI encoding for bold yellow text

val string_of_log : log_type -> string

Get the string representation of a log_type

val fatal : (int * string) -> ('a, unit, string, 'b) Stdlib.format4 -> 'a

A fatal log statement that immediately exits the program

val _log : log_type -> ('a, unit, string, unit) Stdlib.format4 -> 'a

Prints log statements to stdout/stderr

val err : ('a, unit, string, unit) Stdlib.format4 -> 'a