Module Utils.Opengl_utils

Utilities for OpenGL

val bigarray_create : ('a, 'b) Stdlib__Bigarray.kind -> int -> ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

Create a 1D Bigarray of kind k and length len with C layout

parameter k

parameter len

val get_int : ((int32, Stdlib.Bigarray.int32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> 'a) -> int

Get an integer result from OpenGL

parameter f
Operation to get result from

returns
Integer result of running f

val get_string : int -> ((char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> 'a) -> string

Get a string result from OpenGL

parameter len
Length of string to read

parameter f
Operation to get result from

returns
String result of running f

val compile_shaders : string -> string -> int

Compile vertex and fragment shaders into a shader program

parameter vert_src
Vertex shader source code

parameter frag_src
Fragment shader source code

returns
Shader program