Module Utils.Sdl_utils

Utilities for SDL

val let* : 'a Tsdl.Sdl.result -> ('a -> 'b) -> 'b

Let binding for SDL results

val render_draw_circle : Tsdl.Sdl.renderer -> (int * int) -> int -> unit

Draw a circle at center (x, y) with radius https://gist.github.com/Gumichan01/332c26f6197a432db91cc4327fcabb1c

parameter renderer
SDL renderer to do draw

parameter x

parameter y

parameter radius

val render_fill_circle : Tsdl.Sdl.renderer -> (int * int) -> int -> unit

Fill a circle at center (x, y) with radius https://gist.github.com/Gumichan01/332c26f6197a432db91cc4327fcabb1c

parameter renderer
SDL renderer to do draw

parameter x

parameter y

parameter radius

val int_of_hex_string : string -> int

Parse a hex string into an integer

val rgb_of_hex : string -> int * int * int

Get RGB values of a hex code color string

val set_render_color : (int * int * int) -> Tsdl.Sdl.renderer -> unit

Set the renderer draw color

parameter r

parameter g

parameter b

parameter renderer
The renderer to set the color of

val sdlcolor_of_tuple : (int * int * int) -> Tsdl.Sdl.color

Get an SDL Color object from RGB values

val clear_texture_cache : ('a, Tsdl.Sdl.texture) Stdlib.Hashtbl.t -> unit

Clear a cache with SDL textures as the values and destroying the textures

val get_rect_bounding_box : Tsdl.Sdl.rect -> int * int * int * int

Get an Sdl.rect's bounding box

returns
x, y, w, h

val draw_rect_outer_thickness : Tsdl.Sdl.renderer -> Tsdl.Sdl.rect -> int -> unit

Draw a rect with thickness that starts from the provided Sdl.rect and grows outwards

parameter renderer
Application's SDL renderer

parameter base_rect
The rect defining where to draw

parameter thickness
Number of pixels outward to draw