Module World.Grid

World grid logic

val world_width : int

val world_height : int

type world_grid = {

altitude : int array;

event : unit array;

magma : unit array;

water_temp : unit array;

water_current : unit array;

air_temp : unit array;

air_current : unit array;

rain : unit array;

biome : Biomes.biome_tile array;

life : unit array;

civilization : unit array;

}

Collection of 1D array representations of the world_widthxworld_height world grid

Note: unit array denotes an unimplemented feature

val grid : world_grid

World grid object

type world_tile_attr_getter = [

| `Altitude

| `Event

| `Magma

| `WaterTemp

| `WaterCurrent

| `AirTemp

| `AirCurrent

| `Rain

| `Biome

| `Life

| `Civilization

]

Denoting getters for world tile attributes

type world_tile_attr_setter = [

| `AirCurrent of unit

| `AirTemp of unit

| `Altitude of int

| `Biome of Biomes.biome_tile

| `Civilization of unit

| `Event of unit

| `Life of unit

| `Magma of unit

| `Rain of unit

| `WaterCurrent of unit

| `WaterTemp of unit

]

Denoting setters for world tile attributes

val get_grid_attr : int -> world_tile_attr_getter -> world_tile_attr_setter

Get a grid attribute

parameter i
Index into 1D array denoted by attr

parameter attr
Attribute getter

val set_grid_attr : int -> world_tile_attr_setter -> unit

Set a grid attribute

parameter i
Index into 1D array denoted by attr

parameter attr
Attribute setter

val get_grid_tile : ?wrap_x:bool -> (int * int) -> world_tile_attr_getter list -> world_tile_attr_setter list option

Get a list of tile attributes for a world tile

parameter wrap_x
Whether to wrap if x is out of bounds

parameter x
x position of tile

parameter y
y position of tile

parameter fields
List denoting which attributes to get

returns
None if there is no tile at (x, y), otherwise a list of retrieved fields for the tile

val set_grid_tile : ?wrap_x:bool -> (int * int) -> world_tile_attr_setter list -> unit

Set tile attributes for a world tile, or do nothing if there is no tile at (x, y)

parameter wrap_x
Whether to wrap if x is out of bounds

parameter x
x position of tile

parameter y
y position of tile

parameter fields
List denoting which attributes to set