Module Life.Lifeform

Logic for handling the available species of life and their properties

type species =

| Amoeba

| Amphibian

| Bacteria

| Bird

| Carnifern

| Crab

| Dinosaur

| Fish

| Insect

| Mammal

| Octopus

| Reptile

| Robot

| Starfish

| Trichordate

| Whale

Life species

type stage =

| Early

| Middle

| Late

| Final

Evolution stage for a species

type variant = int

Visual species variant selector

type lifeform = {

species : species;

stage : stage;

variant : variant;

}

Describes a single lifeform entity

val string_of_species : species -> string

Get the string representation of a species

val string_of_stage : stage -> string

Get the string representation of an evolution stage

val string_of_variant : variant -> string

Get the string representation of a species variant (for debugging)

val string_of_lifeform : lifeform -> string

Get the string representation of a lifeform (for debugging)

val lifeform_colorkey_rgb : int * int * int

Lifeform sprite transparency color

val blob_of_lifeform : int -> lifeform -> Assets.Assetloader.asset_blob

Get the sprite blob for a lifeform

parameter frame_count
The current frame counter

parameter lf
The lifeform to retrieve a sprite blob for

returns
The sprite blob associated with the lifeform's species, stage, and variant