Module World.World_setup

World generation

val classify_ocean_tile : int -> Biomes.biome_tile

Classify a tile between Biomes.biome_tile.Ocean and Biomes.biome_tile.Land based on altitude

parameter alt
Altitude

returns
Biome tile distinguishing between land and ocean

val latitudinal_falloff : int -> float

Compute a coefficient that scales down as y grows farther from the equator

parameter y
Latitude value

returns
1 - (2 \cdot \frac{\texttt{y}}{height - 1} - 1)^4

val world_setup : unit -> unit

Procedurally generate world grid using simplex noise and FBM