Module Life.Lifeform

Logic for handling the available classes of life and their properties

type life_class =

| Eukaryote

| Amphibian

| Prokaryote

| Avian

| Carnifern

| Arthropod

| Dinosaur

| Fish

| Insect

| Mammal

| Mollusk

| Reptile

| Robot

| Radiate

| Trichordate

| Cetacean

Class of life

type species = int

Class variant selector

type lifeform = {

life_class : life_class;

species : species;

}

Describes a single lifeform entity

val string_of_class : life_class -> string

Get the string representation of a class

val string_of_species : species -> string

Get the string representation of a species variant (for debugging)

val string_of_lifeform : lifeform -> string

Get the string representation of a lifeform (for debugging)

val blob_of_lifeform : int -> lifeform -> Assets.Assetloader.asset_blob

Get the sprite blob for a lifeform

parameter frame_count
The current frame counter

parameter lf
The lifeform to retrieve a sprite blob for

returns
The sprite blob associated with the lifeform's species, stage, and variant

val evolution_step : lifeform -> life_class option

Determine the class of life that a lifeform can evolve into

parameter lf
The lifeform

returns
None if the lifeform cannot evolve, otherwise its potential next class