Module World.World_setup

World generation

val random_lifeform : unit -> Life.Lifeform.lifeform

Generate a completely random lifeform

val latitudinal_falloff : int -> float

Compute a coefficient that scales down as y grows farther from the equator

parameter y
Latitude value

returns
1 - (2 \cdot \frac{\texttt{y}}{height - 1} - 1)^4

val world_setup : unit -> unit

Procedurally generate world grid using simplex noise and FBM